Refactor: Extract database layer into storage.py and fix confirm.html clipboard
