+++ snappass/main.py
@@ -1,15 +1,17 @@
 import os
-import sys
-import secrets
 import typing
 
-import redis
 from flask import (
     abort, Flask, render_template, request, jsonify, make_response,
     Response, Request
 )
-from redis.exceptions import ConnectionError
 from urllib.parse import quote_plus, unquote_plus, urljoin, urlsplit
+from snappass.storage import (
+    check_redis_alive,
+    get_password,
+    password_exists,
+    set_password,
+)
 # _ is required to get the Jinja templates translated
 from flask_babel import Babel, _  # type: ignore # noqa: F401
 
@@ -40,24 +42,6 @@ def get_locale() -> typing.Optional[str]:
 
 babel = Babel(app, locale_selector=get_locale)
 
-# Initialize Redis
-redis_client: redis.StrictRedis
-if os.environ.get('MOCK_REDIS'):
-    from fakeredis import FakeStrictRedis
-
-    redis_client = FakeStrictRedis(version=(6, 2), protocol=2)  # type: ignore
-elif os.environ.get('REDIS_URL'):
-    redis_url = os.environ.get('REDIS_URL')
-    if not redis_url:
-        raise ValueError("REDIS_URL is empty")
-    redis_client = redis.StrictRedis.from_url(redis_url)
-else:
-    redis_host = os.environ.get('REDIS_HOST', 'localhost')
-    redis_port = int(os.environ.get('REDIS_PORT', 6379))
-    redis_db = int(os.environ.get('SNAPPASS_REDIS_DB', 0))
-    redis_client = redis.StrictRedis(
-        host=redis_host, port=redis_port, db=redis_db)
-REDIS_PREFIX: str = os.environ.get('REDIS_PREFIX', 'snappass')
 
 TIME_CONVERSION: typing.Dict[str, int] = {
     'two weeks': 1209600,
@@ -80,22 +64,6 @@ def _request_has_trusted_host(req: Request) -> bool:
         return False
     normalized_hostname = hostname.lower().rstrip('.')
     return normalized_hostname in {'localhost', '127.0.0.1', '::1'}
-
-
-def check_redis_alive(fn: typing.Callable) -> typing.Callable:
-    def inner(*args: typing.Any, **kwargs: typing.Any) -> typing.Any:
-        try:
-            if fn.__name__ == 'main':
-                redis_client.ping()
-            return fn(*args, **kwargs)
-        except ConnectionError as e:
-            print(f'Failed to connect to redis! {e}')
-            if fn.__name__ == 'main':
-                sys.exit(0)
-            else:
-                return abort(500)
-
-    return inner
 
 
 def as_validation_problem(
@@ -140,35 +108,6 @@ def as_problem_response(
     response = make_response(jsonify(problem), status_code)
     response.headers['Content-Type'] = 'application/problem+json'
     return response
-
-
-@check_redis_alive
-def set_password(password: str, ttl: int) -> str:
-    """
-    Store the encrypted password (payload) for the specified lifetime.
-
-    Returns the storage key where the password is stored.
-    """
-    storage_key = f"{REDIS_PREFIX}{secrets.token_urlsafe(16)}"
-    redis_client.set(storage_key, password, ex=ttl)
-    return storage_key
-
-
-@check_redis_alive
-def get_password(storage_key: str) -> typing.Optional[str]:
-    """
-    From a given storage key, return the stored password payload.
-    """
-    password = redis_client.getdel(storage_key)
-
-    if password is not None:
-        return password.decode('utf-8')
-    return None
-
-
-@check_redis_alive
-def password_exists(storage_key: str) -> bool:
-    return bool(redis_client.exists(storage_key))
 
 
 def empty(value: typing.Optional[str]) -> bool:
